requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.2,<3.16
psycopg2>=2.9.10,<2.10
drf-spectacular>=0.28.0,<0.29.0
djangorestframework-simplejwt>=5.5.0,<5.6
django-filter>=25.1,<26.0
